Moves main screen to its own file, adds in option to quit gracefully, moves updating main screen to the main screen
